+++ Source/SwiftLintFramework/Rules/ControlStatementRule.swift
@@ -62,7 +62,7 @@ public struct ControlStatementRule: Rule {
                     return nil
                 }
                 return StyleViolation(ruleDescription: self.dynamicType.description,
-                    location: Location(file: file, offset: match.location))
+                    location: Location(file: file, characterOffset: match.location))
             }
         }