+++ Source/SwiftLintFramework/Rules/LegacyConstructorRule.swift
@@ -37,7 +37,7 @@ public struct LegacyConstructorRule: Rule {
 
         return file.matchPattern(pattern, withSyntaxKinds: [.Identifier]).map {
             StyleViolation(ruleDescription: self.dynamicType.description,
-                location: Location(file: file, offset: $0.location))
+                location: Location(file: file, characterOffset: $0.location))
         }
     }
 }